qa/evals/deep-evals/rag-evals/requirements.txt
langchain
langchain-openai
python-dotenv
langchain-core
langchain-community
faiss-cpu
tiktoken
deepeval